+++ lib/lazy_loading_list_view.dart
@@ -8,7 +8,8 @@ import 'package:shimmer/shimmer.dart';
 typedef ItemLoader<T> = Future<List<T>> Function(int page);
 
 /// Defines a function that takes a BuildContext and an item of type T, and returns a Widget of type T
-typedef ItemBuilder<T> = Widget Function(BuildContext context, T item);
+typedef ItemBuilder<T> = Widget Function(
+    BuildContext context, T item, int index);
 
 /// Defines a function that takes a BuildContext and returns a Widget for Shimmer loading effect
 typedef ShimmerBuilder = Widget Function(BuildContext context);
@@ -24,12 +25,15 @@ class LazyLoadingListView<T> extends StatefulWidget {
   /// Function to build the shimmer effect
   final ShimmerBuilder? shimmerBuilder;
 
+  final int pageSize;
+
   /// Constructor for the widget
   const LazyLoadingListView({
     Key? key,
     required this.loadItems,
     required this.buildItem,
     this.shimmerBuilder,
+    this.pageSize = 10,
   }) : super(key: key);
 
   @override
@@ -44,30 +48,47 @@ class _LazyLoadingListViewState<T> extends State<LazyLoadingListView<T>> {
   /// Boolean to keep track if it's currently loading more items
   bool _loading = false;
 
+  /// Boolean to keep track if there's more items to load
+  bool _canLoadMore = true;
+
   /// The current page number
   int _currentPage = 0;
+
+  /// The scroll controller used to detect if user reaches bottom of ListView
+  final ScrollController _scrollController = ScrollController();
 
   /// Override the initState method to load more items at the start
   @override
   void initState() {
     super.initState();
+    _scrollController.addListener(_onScroll);
     _loadMore();
+  }
+
+  /// Detects if user scrolls to bottom of ListView and loads more data
+  void _onScroll() {
+    if (_scrollController.position.pixels >
+        _scrollController.position.maxScrollExtent - 200) {
+      _loadMore();
+    }
   }
 
   /// Function to load more items
   void _loadMore() {
     /// Only load more items if it's not currently loading
-    if (!_loading) {
+    if (!_loading && _canLoadMore) {
       setState(() {
         _loading = true;
       });
 
       /// Load the items and then update the state
-      widget.loadItems(_currentPage + 1).then((newItems) {
+      widget.loadItems(_currentPage).then((newItems) {
         setState(() {
           _items.addAll(newItems);
           _currentPage++;
           _loading = false;
+          _canLoadMore = newItems.length == widget.pageSize;
+          _onScroll();
         });
       });
     }
@@ -86,6 +107,8 @@ class _LazyLoadingListViewState<T> extends State<LazyLoadingListView<T>> {
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
+      physics: const AlwaysScrollableScrollPhysics(),
+      controller: _scrollController,
       itemCount: _items.length + (_loading ? 1 : 0),
       itemBuilder: (context, index) {
         /// Load more items if it reached the end of the list
@@ -99,8 +122,14 @@ class _LazyLoadingListViewState<T> extends State<LazyLoadingListView<T>> {
         }
 
         /// Build each item using the provided function
-        return widget.buildItem(context, _items[index]);
+        return widget.buildItem(context, _items[index], index);
       },
     );
+  }
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
   }
 }