+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
 Provides ability to lazy load and paginate data within a ListView.
+
+## 0.0.2
+
+You can now specify a pageSize, defaults to 10.
+The buildItem function now provides the current item index.
+The LazyLoadingListView now uses a scrollController to detect when user reaches bottom of ListView.
+LazyLoadingListView now detects when there is no more data to load.
+